♻️ MainPage에서 CourseFilteringTab 별도 컴포넌트로 분리
