+++ src/features/course/course-filtering-tab/index.ts
@@ -1,0 +1,3 @@
+import CourseFilteringTab from './CourseFilteringTab';
+
+export default CourseFilteringTab;
