+++ src/features/course/courses-container/CoursesContainer.tsx
@@ -11,8 +11,9 @@ interface PropsType {
 const CoursesContainer = ({ courses }: PropsType) => {
     return (
         <div className={S.layout}>
-            {courses.map(({ name, memberCount, pinCount, isMyCourse }) => (
+            {courses.map(({ id, name, memberCount, pinCount, isMyCourse }) => (
                 <CourseTab
+                    key={id}
                     name={name}
                     memberCount={memberCount}
                     pinCount={pinCount}
