+++ src/pages/main/MainPage.css.ts
@@ -18,30 +18,6 @@ export const map = style({
     position: 'relative',
 });
 
-export const courseTypeTabsContainer = style({
-    display: 'flex',
-    alignItems: 'flex-start',
-    gap: '24px',
-});
-
-export const courseTypeTab = style([
-    sprinkles({ typography: 'SemiBold20' }),
-    {
-        color: COLOR.Gray400,
-        padding: '0 0 8px 0',
-    },
-]);
-
-export const selectedCourseTypeTab = style([
-    courseTypeTab,
-    sprinkles({ typography: 'ExtraBold20' }),
-    {
-        color: COLOR.Gray900,
-
-        borderBottom: `2px solid ${COLOR.Gray900}`,
-    },
-]);
-
 export const addNewCourseButton = style({
     display: 'flex',
     alignItems: 'center',
